refactor: extract manifest trust store helpers
